requirements.txt
pydantic-ai
python-dotenv==1.0.0
langgraph
langchain-anthropic
langchain-openai
autogen-agentchat~=0.2
matplotlib
numpy
langchain
duckduckgo-search
langchain-google-genai
langchain_community
flask==3.0.0
flask-cors
flask-restful
flask-limiter
flask-sqlalchemy
streamlit>=1.24.0
gunicorn==21.2.0
requests>=2.31.0
tweepy